@@ -1,41 +1,29 @@
-name: sistemarenascerdaesperanca
-description: A new Flutter project.
-publish_to: "none" # Remove this line if you wish to publish to pub.dev
-
+name: renascer_sistema
+description: "A new Flutter project."
+publish_to: 'none' # Remove this line if you wish to publish to pub.dev
 version: 1.0.0+1
 
 environment:
-  sdk: ">=3.1.5 <4.0.0"
+  sdk: ^3.5.4
 
 dependencies:
   flutter:
     sdk: flutter
-  cupertino_icons: ^1.0.2
-  firebase_auth: ^4.15.0
-  firebase_core: ^2.24.0
-  firebase_storage: ^11.6.5
-  cloud_firestore: ^4.13.3
-  http: ^1.1.0
-  provider: ^6.1.1
-  flutter_modular: ^6.3.2
-  file_picker: ^8.0.0+1
-  uuid: ^4.4.0
-  image_picker: ^1.1.2
+
+  cupertino_icons: ^1.0.8
+  firebase_core: ^3.8.1
+  firebase_auth: ^5.3.4
+  cloud_firestore: ^5.5.1
+  firebase_storage: ^12.3.7
+  # file_picker: ^8.1.5
+  # image_picker: ^1.1.2
 
 dev_dependencies:
   flutter_test:
     sdk: flutter
-  flutter_lints: ^2.0.0
 
+  flutter_lints: ^4.0.0
+
+# The following section is specific to Flutter packages.
 flutter:
   uses-material-design: true
-  assets:
-    - assets/images/logo.jpg
-
-fonts:
-  - family: roboto
-    fonts:
-      - asset: assets/fonts/Roboto-Medium.ttf
-        weight: 500
-      - asset: assets/fonts/Roboto-Regular.ttf
-        weight: 400
