@@ -3,14 +3,14 @@ import 'package:flutter/material.dart';
 class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
   final Function() onMenuPressed;
 
-  const CustomAppBar({Key? key, required this.onMenuPressed}) : super(key: key);
+  const CustomAppBar({super.key, required this.onMenuPressed, required bool showFilterIcon, required Text leftContent});
 
   @override
   Widget build(BuildContext context) {
     return AppBar(
       title: const Text('Renascer Da Esperança'),
       leading: IconButton(
-        icon: Icon(Icons.menu),
+        icon: const Icon(Icons.menu),
         onPressed: onMenuPressed,
       ),
     );
@@ -23,28 +23,26 @@ class CustomAppBar extends StatelessWidget implements PreferredSizeWidget {
 class SettingsMenu extends StatelessWidget {
   final Function() onSettingSelected;
 
-  const SettingsMenu({Key? key, required this.onSettingSelected}) : super(key: key);
+  const SettingsMenu({super.key, required this.onSettingSelected});
 
   @override
   Widget build(BuildContext context) {
-    return Container(
-      child: Column(
-        mainAxisSize: MainAxisSize.min,
-        children: <Widget>[
-          ListTile(
-            title: const Text('Configuração 1'),
-            onTap: () {
-              onSettingSelected();
-            },
-          ),
-          ListTile(
-            title: const Text('Configuração 2'),
-            onTap: () {
-              onSettingSelected();
-            },
-          ),
-        ],
-      ),
+    return Column(
+      mainAxisSize: MainAxisSize.min,
+      children: <Widget>[
+        ListTile(
+          title: const Text('Configuração 1'),
+          onTap: () {
+            onSettingSelected();
+          },
+        ),
+        ListTile(
+          title: const Text('Configuração 2'),
+          onTap: () {
+            onSettingSelected();
+          },
+        ),
+      ],
     );
   }
 }
