@@ -1,8 +1,19 @@
 // File generated by FlutterFire CLI.
-// ignore_for_file: lines_longer_than_80_chars, avoid_classes_with_only_static_members
+// ignore_for_file: type=lint
 import 'package:firebase_core/firebase_core.dart' show FirebaseOptions;
 import 'package:flutter/foundation.dart'
     show defaultTargetPlatform, kIsWeb, TargetPlatform;
+
+/// Default [FirebaseOptions] for use with your Firebase apps.
+///
+/// Example:
+/// ```dart
+/// import 'firebase_options.dart';
+/// // ...
+/// await Firebase.initializeApp(
+///   options: DefaultFirebaseOptions.currentPlatform,
+/// );
+/// ```
 class DefaultFirebaseOptions {
   static FirebaseOptions get currentPlatform {
     if (kIsWeb) {
@@ -50,7 +61,7 @@ class DefaultFirebaseOptions {
 
   static const FirebaseOptions android = FirebaseOptions(
     apiKey: 'AIzaSyCM5jVryxXv0_t0EkXNnodAYACqhabg9ao',
-    appId: '1:779344555546:android:67abaaabdec6d17a8d2ba3',
+    appId: '1:779344555546:android:675890f306f77f658d2ba3',
     messagingSenderId: '779344555546',
     projectId: 'sistemarenascerdaesperanca',
     storageBucket: 'sistemarenascerdaesperanca.appspot.com',
