@@ -1,8 +1,8 @@
 // ignore_for_file: library_private_types_in_public_api
+import '../../../styles/widget/iput_decoration.dart';
 import 'cadastra_familiar_model.dart';
 import 'package:flutter/material.dart';
 import '../../../styles/colors_app.dart';
-import 'package:sistemarenascerdaesperanca/src/styles/widget/iput_decoration.dart';
 
 class CadastroPessoaPage extends StatefulWidget {
   const CadastroPessoaPage({Key? key}) : super(key: key);
